requirements.txt
alembic==1.14.1
blinker==1.8.2
click==8.1.8
flask==3.0.3
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
MarkupSafe==2.1.5
python-dotenv==1.0.1
sqlalchemy==2.0.40
typing-extensions==4.13.2
werkzeug==3.0.6
wtforms==3.1.2
zipp==3.20.2
